requirements.txt
Django==2.2.28
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.12.4
djoser==2.1.0
Pillow==9.3.0
